Add timeout to uart receiving
Much like on the IO board side, this adds a timeout for when we're
waiting for some data from the IO board. Occasionally (well, with some
regularity) we somehow don't pick up on receipt of data, which keeps us
hanging in a waiting state, locking the whole system up.

Now we have a timeout in those cases; if we don't hear back in a couple
of ticks, we abort the receive DMA request and reset the port_state so
we can move on. This means that glitches that otherwise would hang the
entire system are now quickly recovered from, making everything a lot
more stable (read: actually working)