Better unflickerify LEDs
Brittany figured out the logical reason LEDs were flickering and
devised this fix, setting the timer count 0.

Runnign the ADC and the timer at the same time results in a situation
where the timer's CNT isn't correctly reset to 0.

The timer was sometimes moving to 0 later, causing a late MINC (memory
increment) in the DMA request, and this lateness cascades, causing
corrupted data to be written to the LEDs. Manually ensuring CNT is 0
before starting the LED DMA stuff effectively prevents this issue.